Move /install-confirm to routes folder
