+++ apps/Envoy/routes/index.js
@@ -17,6 +17,7 @@ const { verifyUrlForSlack } = require('./verify-url-for-slack');
 // const { redirect } = require('./redirect');
 const { employeeUpcoming } = require('./employee-upcoming');
 const { upcomingVisit } = require('./upcoming-visit');
+const { installConfirm } = require('./install_confirm');
 
 /** Registers routes for express router for incoming data from Envoy */
 const registerCustomRoutes = () => {
@@ -38,6 +39,7 @@ const registerCustomRoutes = () => {
     // routes.push(redirect);
     routes.push(employeeUpcoming);
     routes.push(upcomingVisit);
+    routes.push(installConfirm);
     return routes;
 };
