+++ slackApp.js
@@ -19,7 +19,6 @@ const { webClientUser, webClientBot } = require('./zArchive/SlackHelper');
 // const { EnvoyAPI, middleware, errorMiddleware, asyncHandler, EnvoyResponseError } = require('@envoy/envoy-integrations-sdk');
 // const request = require('request');  //Change to Axios
 const axios = require('axios');
-const FormData = require('form-data');
 // const Envoy = require('./Envoy');
 // const getAccessToken = require('./getAccessToken');
 // //const envoy-auth = require('./apps/Envoy/middleware/envoy-auth')
@@ -80,52 +79,6 @@ const slackApp = new App(
     },
     logLevel: LogLevel.DEBUG,
     receiver,
-    customRoutes: [
-      {
-        path: '/install-confirm',
-        method: ['GET'],
-        handler: async (req, res) => {
-          try {
-            let begin = req.url.indexOf("code") + 5;
-            let end = req.url.indexOf("&");
-            const code = req.url.slice(begin, end);
-            const form = new FormData();
-            form.append('code', code);
-            form.append('client_id', process.env.SLACK_CLIENT_ID);
-            form.append('client_secret', process.env.SLACK_CLIENT_SECRET);
-
-            const response = await axios.post(
-              'https://slack.com/api/oauth.v2.access',
-              form,
-              {
-                headers: {
-                  ...form.getHeaders()
-                }
-              }
-            );
-
-            const install = response.data;
-            if (!install.is_enterprise_install) {
-              redisClient.hSet(install.team.id, 
-                'id', install.team.id,
-                'name', install.team.name,
-                'tokenType', 'bot',
-                'authVersion', 'v2',
-                'bot_user_id', install.bot_user_id,
-                'botToken', install.access_token,
-              );
-              res.end();
-            } else {
-              // Handle enterprise install.
-            }
-          } catch (e) {
-            console.error(e);
-            res.writeHead(500);
-            res.end('Failed to install', 'utf-8');
-          }
-        }
-      },
-    ],
   }
 );
